# docs/04-api-spec.yaml
openapi: 3.1.0
info:
  title: FlagOps API
  description: Feature Flag & Application Configuration Service
  version: 0.1.0
paths:
...
  /api/v1/environments/{environment_id}/change-requests:
    post:
      tags:
      - Change Requests
      summary: Tạo change request mới cho một environment
      operationId: create_change_request_api_v1_environments__environment_id__change_requests_post
      security:
      - HTTPBearer: []
      parameters:
      - name: environment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Environment Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRequestCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Change Requests
      summary: Danh sách change request trong environment (lọc theo status)
      operationId: list_change_requests_api_v1_environments__environment_id__change_requests_get
      security:
      - HTTPBearer: []
      parameters:
      - name: environment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Environment Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ChangeRequestStatus'
          - type: 'null'
          description: Lọc theo trạng thái change request
          title: Status
        description: Lọc theo trạng thái change request
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChangeRequestResponse'
                title: Response List Change Requests Api V1 Environments  Environment
                  Id  Change Requests Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/change-requests/{change_request_id}:
    get:
      tags:
      - Change Requests
      summary: Chi tiết change request
      operationId: get_change_request_api_v1_change_requests__change_request_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: change_request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Change Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/change-requests/{change_request_id}/impact:
    get:
      tags:
      - Change Requests
      summary: Mô phỏng tác động của change request trên 1000 context gần nhất
      operationId: get_change_request_impact_api_v1_change_requests__change_request_id__impact_get
      security:
      - HTTPBearer: []
      parameters:
      - name: change_request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Change Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestImpactResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/change-requests/{change_request_id}/approve:
    post:
      tags:
      - Change Requests
      summary: Duyệt change request (nguyên tắc bốn mắt, chỉ ADMIN/OWNER)
      operationId: approve_change_request_api_v1_change_requests__change_request_id__approve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: change_request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Change Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/change-requests/{change_request_id}/reject:
    post:
      tags:
      - Change Requests
      summary: Từ chối change request (chỉ ADMIN/OWNER)
      operationId: reject_change_request_api_v1_change_requests__change_request_id__reject_post
      security:
      - HTTPBearer: []
      parameters:
      - name: change_request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Change Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/change-requests/{change_request_id}/cancel:
    post:
      tags:
      - Change Requests
      summary: Hủy change request (người tạo hoặc ADMIN/OWNER)
      operationId: cancel_change_request_api_v1_change_requests__change_request_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: change_request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Change Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChangeRequestCreate:
      properties:
        title:
          type: string
          maxLength: 200
          title: Title
          description: Brief summary of proposed change
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Detailed explanation or rationale
        payload:
          additionalProperties: true
          type: object
          title: Payload
          description: JSON representation of target changes
        scheduled_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Scheduled At
          description: Optional scheduled execution time
        status:
          $ref: '#/components/schemas/ChangeRequestStatus'
          description: Initial status (defaults to PENDING)
          default: PENDING
      type: object
      required:
      - title
      - payload
      title: ChangeRequestCreate
      description: Payload to create a new change request.
    ChangeRequestImpactResponse:
      properties:
        total_contexts:
          type: integer
          title: Total Contexts
          description: Total evaluated historical contexts (up to 1000)
        affected_contexts:
          type: integer
          title: Affected Contexts
          description: Number of contexts whose evaluated variation changed
        change_percentage:
          type: number
          title: Change Percentage
          description: Percentage of contexts affected (0-100%)
        transitions:
          items:
            $ref: '#/components/schemas/ImpactTransition'
          type: array
          title: Transitions
          description: Breakdown of variation transitions
        summary:
          type: string
          title: Summary
          description: Human-readable impact explanation
        flag_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Flag Key
          description: Key of the primary flag simulated
      type: object
      required:
      - total_contexts
      - affected_contexts
      - change_percentage
      - summary
      title: ChangeRequestImpactResponse
      description: Impact simulation result comparing current vs simulated rulesets.
    ChangeRequestResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        environment_id:
          type: string
          format: uuid
          title: Environment Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        payload:
          additionalProperties: true
          type: object
          title: Payload
        status:
          $ref: '#/components/schemas/ChangeRequestStatus'
        requested_by:
          type: string
          format: uuid
          title: Requested By
        reviewed_by:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Reviewed By
        scheduled_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Scheduled At
        applied_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Applied At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - environment_id
      - title
      - payload
      - status
      - requested_by
      - created_at
      - updated_at
      title: ChangeRequestResponse
      description: Change request details response.
    ChangeRequestStatus:
      type: string
      enum:
      - DRAFT
      - PENDING
      - APPROVED
      - REJECTED
      - APPLIED
      - CANCELLED
      title: ChangeRequestStatus
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ImpactTransition:
      properties:
        from_variation:
          type: string
          title: From Variation
          description: Previous variation name or key
        to_variation:
          type: string
          title: To Variation
          description: Simulated variation name or key
        count:
          type: integer
          title: Count
          description: Number of evaluated contexts undergoing this transition
      type: object
      required:
      - from_variation
      - to_variation
      - count
      title: ImpactTransition
      description: Single variation transition record in impact simulation.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer